言語一覧を表示する機能を追加

+++ commands.py
@@ -1,6 +1,6 @@
 import discord
 from discord.ext import commands
-from gtts import gTTS
+from gtts import gTTS ,lang
 import asyncio
 import os
 
@@ -18,7 +18,7 @@ def setup(bot):
         help_message = "```"
         sorted_commands = sorted(bot.commands, key=lambda x: x.name)
         for command in sorted_commands:
-            help_message += f"!{command.name}: {command.description}\n"
+            help_message += f"!{command.name}: {command.description}\n\n"
         help_message += "```"
         embed = discord.Embed(title="コマンド一覧(アルファベット順)" ,color=0x00ff00 ,description=help_message)
         await ctx.send(embed=embed)
@@ -66,8 +66,8 @@ def setup(bot):
         await ctx.send(f"{member.mention}のミュートを解除しました")
 
     class GTTSEngine:
-        def save_speech(self, text, path):
-            tts = gTTS(text=text, lang='ja', slow=False)  # 日本語で音声合成
+        def save_speech(self, text, lang_code, path):
+            tts = gTTS(text=text, lang=lang_code, slow=False)
             tts.save(path)
 
     async def play_speech(voice_client: discord.VoiceClient, path):
@@ -77,17 +77,33 @@ def setup(bot):
         while voice_client.is_playing():
             await asyncio.sleep(1)
 
-    @bot.command(description="ボイスチャット内で喋ります")
-    async def say(ctx, *, message):
-        if ctx.voice_client is None:
+    @bot.command(description="ボイスチャンネル内で言語に応じて喋ります")
+    async def say(ctx, lang_code: str, *, message: str):
+        was_connected = ctx.voice_client is not None  # ボットが既に接続されているかをチェック
+        supported_langs = lang.tts_langs()  # gttsがサポートする言語のリストを取得
+
+        if not was_connected:
             if ctx.author.voice:
                 await ctx.author.voice.channel.connect()
             else:
                 await ctx.send("ボイスチャンネルに参加してから再度試してください。")
                 return
 
+        if lang_code not in supported_langs:
+            await ctx.send("サポートされていない言語です。")
+            return
+
         tts = GTTSEngine()
-        path = "speech.mp3"  # 音声ファイルの保存パス
-        tts.save_speech(message, path)
+        path = f"speech_{lang_code}.mp3"  # 言語ごとにファイル名を設定
+        tts.save_speech(message, lang_code, path)
         await play_speech(ctx.voice_client, path)
         os.remove(path)  # 再生後にファイルを削除
+
+        if not was_connected:
+            await ctx.voice_client.disconnect()  # ボットが元々接続されていなかった場合、切断します
+
+    
+    @bot.command(description="対応言語の一覧を表示します")
+    async def langlist(ctx):
+        with open('languagelist.txt', 'rb') as file:
+            await ctx.send("対応言語一覧\nSupported Languages", file=discord.File(file, 'languagelist.txt'))
